+++ resources/views/admin/newnote.blade.php
@@ -22,8 +22,7 @@ New Note « Admin CP
 @stop
 
 @section('scripts')
-<link rel="styelsheet" href="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.css">
-<script src="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.js"></script>
+@include('templates.mapbox-links')
 
 <script src="/assets/js/newnote.js"></script>
 <script src="/assets/bower/store2.min.js"></script>