+++ resources/views/admin/newplace.blade.php
@@ -19,8 +19,7 @@ New Place « Admin CP
 @stop
 
 @section('scripts')
-<link rel="stylesheet" href="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.css">
-<script src="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.js"></script>
+@include('templates.mapbox-links')
 
 <script src="/assets/js/newplace.js"></script>
 @stop