+++ resources/views/micropubnewnotepage.blade.php
@@ -32,8 +32,7 @@ New Note « Jonny Barnes
 @stop
 
 @section('scripts')
-<link rel="stylesheet" href="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.css">
-<script src="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.js"></script>
+@include('templates.mapbox-links')
 
 <script src="/assets/bower/fetch.js"></script>
 <script src="/assets/bower/store2.min.js"></script>